Renamed util/views -> util/view_utils
